Rename Strategies => Idempotency
